2014-10-08 02:00 UTC+0200 Przemyslaw Czerpak (druzus/at/poczta.onet.pl)   * contrib/hbnetio/netiosrv.c   * src/rtl/gtpca/gtpca.c     * pacified warnings
  * src/rtl/hbjson.c
    ! include hbjson.h

+++ src/rtl/gtpca/gtpca.c
@@ -279,7 +279,7 @@ static void hb_gt_pca_AnsiGetCurPos( int * iRow, int * iCol )
                continue;
             }
          }
-         if( n == sizeof( rdbuf ) )
+         if( n == ( int ) sizeof( rdbuf ) )
             break;
          time = hb_dateMilliSeconds();
          if( time > end_timer )
